pyproject.toml
[tool.poetry]
name = "quantlab"
version = "1.0.0"
description = "A quantitative trading lab for cryptocurrency futures"
authors = ["QuantLab Team"]
readme = "README.md"
packages = [{include = "quantlab"}]

[tool.poetry.dependencies]
python = "^3.11"
ccxt = "^4.1.0"
pandas = "^2.1.0"
numpy = "^1.24.0"
matplotlib = "^3.7.0"
mplfinance = "^0.12.0"
ta = "^0.10.2"
pydantic = "^2.5.0"
tqdm = "^4.66.0"
pyyaml = "^6.0.1"
click = "^8.1.0"
numba = "^0.58.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
